updated by Srishailam

+++ partyanalyst-service/trunk/src/main/java/com/itgrids/partyanalyst/dao/IPublicRepresentativeDAO.java
@@ -7,6 +7,6 @@ import org.appfuse.dao.GenericDao;
 import com.itgrids.partyanalyst.model.PublicRepresentative;
 
 public interface IPublicRepresentativeDAO extends GenericDao<PublicRepresentative, Long> {
-	public List<Long> getRepresentativesByPositions(Long representativeLevelId,List<Long> locationValuesList,Long positionId);
+	public List<Integer> getRepresentativesByPositions(Long representativeLevelId,List<Long> locationValuesList,Long positionId);
 	public List<Object[]> getCandidateInfoByCandidateIds(List<Long> candidateIdsList);
 }
